+++ src/ECS/mod.rs
@@ -160,7 +160,7 @@ mod tests{
             "gmSys_movement"
         }
     
-        fn execute(&mut self, IN_data: Self::sysData) {
+        fn execute(&mut self, _IN_data: Self::sysData) {
             todo!()
         }
     }
