--- src/manufacture/init.rs
@@ -1,1 +1,0 @@
-use super::*;
