+++ src/manufacture/mod.rs
@@ -8,5 +8,6 @@ pub mod components;
 pub mod resources;
 pub mod systems;
 pub mod vars;
+pub mod types;
 
 pub mod init;
