+++ src/manufacture/resources/event.rs
@@ -1,7 +1,7 @@
 use super::*;
 
 use std::rc::Rc;
-use std::cell::{Ref, RefCell, RefMut};
+use std::cell::RefCell;
 
 pub struct res_Events{
     pub activeBuffer: bool,
